Refactor: DTO 패키지를 도메인 기준으로 기능별 분리
- teum, schedule, availability, shared, common 하위 패키지 생성
- 기능별로 DTO 그룹 구조 정리 (요청/응답 혼합 유지)
- 로직 변경 없음
