+++ src/main/java/umc/teumteum/server/domain/teum/dto/schedule/ScheduledTeumDetailResponseDto.java
@@ -1,10 +1,11 @@
-package umc.teumteum.server.domain.teum.dto;
+package umc.teumteum.server.domain.teum.dto.schedule;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import umc.teumteum.server.domain.teum.dto.common.ParticipantDto;
 
 import java.util.List;
