+++ src/main/java/umc/teumteum/server/domain/teum/dto/schedule/ScheduledTeumExitResponseDto.java
@@ -1,4 +1,4 @@
-package umc.teumteum.server.domain.teum.dto;
+package umc.teumteum.server.domain.teum.dto.schedule;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
