+++ src/main/java/umc/teumteum/server/domain/teum/dto/teum/TeumResponseStatus.java
@@ -1,4 +1,4 @@
-package umc.teumteum.server.domain.teum.dto;
+package umc.teumteum.server.domain.teum.dto.teum;
 
 public enum TeumResponseStatus {
     ACCEPTED,
